requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
chardet==3.0.4
colorama==0.4.6
exceptiongroup==1.3.0
execnet==2.1.1
idna==2.10
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-rerunfailures==15.1
pytest-xdist==3.8.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.23.0
rpds-py==0.26.0
soupsieve==2.7
tomli==2.2.1
typing_extensions==4.14.1
urllib3==1.25.11
